+++ seedimg-png/seedimg-png.cpp
@@ -119,7 +119,8 @@ seedimg::modules::png::from(const std::string &filename) noexcept {
   if (interlace_type == PNG_INTERLACE_NONE) {
     for (std::size_t y = 0; y < res_img->height; ++y) {
       png_read_row(png_ptr,
-                   reinterpret_cast<png_bytep>(res_img->get_row(y).data()),
+                   reinterpret_cast<png_bytep>(
+                       res_img->get_row(static_cast<uint32_t>(y)).data()),
                    nullptr);
     }
   } else {
@@ -133,7 +134,9 @@ seedimg::modules::png::from(const std::string &filename) noexcept {
       png_bytep row = row_pointers[y];
       for (std::size_t x = 0; x < res_img->width; ++x) {
         png_bytep px = &(row[x * 4]);
-        res_img->get_pixel(x, y) = {px[0], px[1], px[2], px[3]};
+        res_img->get_pixel(static_cast<uint32_t>(x),
+                           static_cast<uint32_t>(y)) = {px[0], px[1], px[2],
+                                                        px[3]};
       }
       free(row);
     }
@@ -205,7 +208,8 @@ bool seedimg::modules::png::to(
 
   for (std::size_t y = 0; y < inp_img->height; ++y) {
     png_write_row(png_ptr,
-                  reinterpret_cast<png_bytep>(inp_img->get_row(y).data()));
+                  reinterpret_cast<png_bytep>(
+                      inp_img->get_row(static_cast<uint32_t>(y)).data()));
   }
 
   png_write_end(png_ptr, nullptr);
